backend/requirements.txt
fastapi
uvicorn
spacy
rapidfuzz
pymupdf
python-multipart
pydantic
httpx
python-dotenv
